refactor(admin): consolidate agent and role-service management module
Pages for configuring agents, role/model-service bindings and the shared useAgents logic were split across view/, composables/ and styles/. Move AgentConfig.vue, RoleService.vue, useAgents.ts and agents.scss under admin/agents so the agent management feature stays cohesive.
