+++ webui/src/admin/agents/agents.scss
